tg-bot/bot/requirements.txt
python-telegram-bot
asyncio
aiohttp